# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vo de Taxi",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cars": {
            "get": {
                "tags": [
                    "#3 - Car"
                ],
                "summary": "Display all cars",
                "description": "Display all cars",
                "operationId": "display_cars",
                "responses": {
                    "201": {
                        "description": "Clients registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            },
            "post": {
                "tags": [
                    "#3 - Car"
                ],
                "summary": "Register a new car",
                "description": "Register a new car",
                "operationId": "register_car",
                "parameters": [
                    {
                        "name": "driver_id",
                        "in": "path",
                        "description": "Driver driver_id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "plate",
                                    "color",
                                    "year",
                                    "model",
                                    "name"
                                ],
                                "properties": {
                                    "plate": {
                                        "type": "text"
                                    },
                                    "color": {
                                        "type": "text"
                                    },
                                    "year": {
                                        "type": "number"
                                    },
                                    "model": {
                                        "type": "text"
                                    },
                                    "name": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "plate": "FAB-8575",
                                        "color": "black",
                                        "year": "2013",
                                        "model": "FIAT",
                                        "name": "BRAVO"
                                    }
                                ]
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/cars/{id}": {
            "put": {
                "tags": [
                    "#3 - Car"
                ],
                "summary": "Update a car",
                "description": "Update a car",
                "operationId": "update_car",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Card id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "plate",
                                    "color",
                                    "year",
                                    "model",
                                    "name"
                                ],
                                "properties": {
                                    "plate": {
                                        "type": "text"
                                    },
                                    "color": {
                                        "type": "text"
                                    },
                                    "year": {
                                        "type": "number"
                                    },
                                    "model": {
                                        "type": "text"
                                    },
                                    "name": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "plate": "FAB-8575",
                                        "color": "black",
                                        "year": "2013",
                                        "model": "FIAT",
                                        "name": "BRAVO"
                                    }
                                ]
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}